+++ src/main/scala/com/scalamandra/model/dto/auth/ActivationRequest.scala
@@ -1,0 +1,6 @@
+package com.scalamandra.model.dto.auth
+
+case class ActivationRequest(
+                              token: String,
+                              userId: Long,
+                            )
