+++ src/main/scala/com/scalamandra/service/AuthService.scala
@@ -1,7 +1,7 @@
 package com.scalamandra.service
 
-import com.scalamandra.model.HttpException.{Conflict, Unauthorized}
-import com.scalamandra.model.dto.{AuthedUser, LoginRequest, LoginResponse, RefreshRequest, RefreshResponse, RegisterRequest}
+import com.scalamandra.model.HttpException.{Conflict, NotFound, Unauthorized}
+import com.scalamandra.model.dto.auth._
 
 import scala.concurrent.Future
 
@@ -12,5 +12,7 @@ trait AuthService extends Service {
   def login(request: LoginRequest): Future[Either[Unauthorized, LoginResponse]]
 
   def refresh(authedUser: AuthedUser, request: RefreshRequest): Future[Either[Unauthorized, RefreshResponse]]
+
+  def activate(request: ActivationRequest): Future[Either[NotFound, Unit]]
 
 }
